Prep for taking out file writer engine and using it in the clock class \n also testing if I can use newline character here
